NZBGet #372: hold the re-armed cloud Lizard gates after the #388 merge
Merge-QA (4-lens Fable workflow) on the #388-into-#373 merge confirmed the
merge semantics are clean (interface drift both directions, #372 invariants,
cross-file merges) but found the PR trips cloud gates #388 just re-armed --
invisible to local gates (pylint too-many-lines is disabled; cloud Lizard's
file-nloc>500 fires only in repo-level analysis after merge):

- nzbget_resolver.py was 736 NLOC (>500; main baseline 408). Split via the
  resolver_fallback_jobs facade idiom into nzbget_resolver_dupes.py (the #372
  Smart-Duplicates backup fleet, 216 NLOC) and nzbget_resolver_smb.py (the
  completed-download SMB mapping + video discovery, 115 NLOC); parent now 443.
  Children route every sibling/parent reference through call-time
  `_core.<name>` so all existing resources.lib.nzbget_resolver.* patch targets
  keep intercepting; every moved name is re-exported.

- router.py was 511 NLOC (>500; main baseline 499). _handle_play's inline
  identity dict hoisted to router_play._play_identity, and
  _PROVIDER_SEARCH_SETTING_DEFAULTS moved to router_search.py; router now 496.

- The defaults move carries the QA finding's fix: hydra_url now seeds the
  schema default in the provider-search snapshot, so a default-URL Hydra setup
  keeps hydra_url-gated features on the snapshotted search path too (the
  snapshot pre-reads every key, which bypassed the hydra._DEFAULT_HYDRA_URL
  mirror). test_search_all_providers_uses_script_settings_getter_without_
  kodi_addon updated to the new contract; new
  test_provider_search_defaults_mirror_hydra_schema_url covers it.

Cloud-gate reproduction on the merged tree: file-nloc>500 none, params>8 none,
ccn>8 only the two pre-existing main-side CCN-9s (accepted, untouched).
just lint 10.00/10; just test 2233 passed / 2 skipped.
