+++ repo/plugin.video.nzbdav/resources/lib/nzbget_resolver.py
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # Copyright (C) 2026 nzbdav contributors
+# pylint: disable=cyclic-import,unused-import
 
 """NZBGet backend resolver: submit to NZBGet, wait, play from SMB.
 
@@ -15,7 +16,6 @@ from urllib.parse import unquote
 import xbmc
 import xbmcgui
 import xbmcplugin
-import xbmcvfs
 
 from resources.lib import nzbget_api
 from resources.lib.download_ledger import record_download
@@ -24,234 +24,41 @@ from resources.lib.http_util import redact_text as _redact_text
 from resources.lib.i18n import addon_name as _addon_name
 from resources.lib.i18n import fmt as _fmt
 from resources.lib.i18n import string as _string
+from resources.lib.nzbget_resolver_dupes import (  # noqa: E402,F401
+    _HEALTHCHECK_LOCK,
+    _HEALTHCHECK_WARNED,
+    _MAX_EXTRA_BACKUPS,
+    _append_one_backup,
+    _cleanup_canceled_submissions,
+    _dupe_check_disabled,
+    _dupe_worker_should_skip,
+    _extra_backups_from_loader,
+    _load_extra_candidates,
+    _snapshot_conn_getter,
+    _spawn_dupe_backups,
+    _submit_backup_fleet,
+    _submit_dupe_backups,
+    _usable_backup_link,
+    _warn_if_healthcheck_pauses,
+)
+from resources.lib.nzbget_resolver_smb import (  # noqa: E402,F401
+    _SMB_LIST_RETRY_INTERVAL,
+    _SMB_MAX_DEPTH,
+    _SMB_RESOLVE_BUDGET,
+    VIDEO_EXTENSIONS,
+    _drive_resolve_dialog,
+    _is_video_name,
+    _largest_video_in_dir,
+    _largest_video_in_tree,
+    _smb_exact_mapping,
+    _smb_fallback_mapping,
+    _smb_file_size,
+    nzbget_smb_target,
+    pick_largest_video,
+    resolve_smb_video,
+)
 
 # Same extensions the WebDAV path uses (webdav.py VIDEO_EXTENSIONS).
-VIDEO_EXTENSIONS = (".mkv", ".mp4", ".avi", ".m4v", ".ts", ".wmv", ".mov")
-
-
-def nzbget_smb_target(smb_root, dest_dir, category="", completed_base=""):
-    """Map NZBGet's server-local DestDir onto the SMB root.
-
-    ``smb_root`` points at NZBGet's *completed* base dir, exposed over SMB.
-
-    Preferred mapping (exact): when ``completed_base`` (NZBGet's configured
-    global completed DestDir) is known and is a prefix of ``dest_dir``, map the
-    *relative* remainder onto ``smb_root``. This mirrors whatever subfolder
-    layout NZBGet actually used — ``AppendCategoryDir`` on/off, a category that
-    matches or a category-specific/custom ``DestDir`` whose folder name differs
-    from the category setting (e.g. ``<completed>/films/<release>`` for category
-    ``movies``) — without guessing.
-
-    Fallback heuristic (``completed_base`` unknown / ``dest_dir`` outside it):
-    derive the category segment from ``dest_dir``'s own parent. Nest under it
-    only when a category is configured and the parent isn't already the SMB
-    root's own trailing segment (the AppendCategoryDir=no case, where the
-    parent is the completed base itself). Returns the smb:// folder URL, or
-    None if dest_dir is empty.
-    """
-    if not dest_dir:
-        return None
-    normalized = dest_dir.replace("\\", "/").rstrip("/")
-    base = smb_root.rstrip("/")
-
-    exact = _smb_exact_mapping(normalized, base, completed_base)
-    if exact is not None:
-        return exact
-
-    return _smb_fallback_mapping(normalized, base, category)
-
-
-def _smb_exact_mapping(normalized, base, completed_base):
-    """Map ``dest_dir`` onto ``base`` via the configured completed base.
-
-    Returns the smb:// URL when ``completed_base`` is known and a strict
-    prefix of ``dest_dir``, else ``None`` so the caller falls back to the
-    heuristic.
-    """
-    cb = (completed_base or "").replace("\\", "/").rstrip("/")
-    if not cb or normalized.casefold() == cb.casefold():
-        return None
-    if not (normalized + "/").casefold().startswith(cb.casefold() + "/"):
-        return None
-    rel = normalized[len(cb) :].strip("/")
-    # Guard against a doubled tail when smb_root was pointed at a
-    # subfolder of the completed base (e.g. root .../movies + rel
-    # movies/Release): drop rel's leading segment if base already ends
-    # with it.
-    rel_head = rel.split("/", 1)[0]
-    if rel_head and base.casefold().endswith("/" + rel_head.casefold()):
-        rel = rel[len(rel_head) :].strip("/")
-    if rel:
-        return "{}/{}".format(base, rel)
-    return None
-
-
-def _smb_fallback_mapping(normalized, base, category):
-    """Derive the SMB target from ``dest_dir``'s own tail segments.
-
-    Used when the completed base is unknown / outside ``dest_dir``. Returns
-    the smb:// URL, or ``None`` when ``dest_dir`` has no usable segments.
-    """
-    segments = list(filter(None, normalized.split("/")))
-    if not segments:
-        return None
-    release_folder = segments[-1]
-    parent_folder = segments[-2] if len(segments) >= 2 else ""
-    category = (category or "").strip().strip("/")
-    # Without the completed base we can't reliably tell a real category
-    # subfolder apart from the completed base's own last segment (the
-    # server-side folder name may differ from the SMB share alias). So only
-    # nest when DestDir's parent is *exactly* the configured category — the
-    # one case we can be sure about. Otherwise map the release folder directly
-    # under the SMB root (correct for AppendCategoryDir=no, and the safe
-    # default for a custom DestDir whose layout we can't confirm).
-    if (
-        category
-        and parent_folder.casefold() == category.casefold()
-        and not base.casefold().endswith("/" + category.casefold())
-    ):
-        return "{}/{}/{}".format(base, parent_folder, release_folder)
-    return "{}/{}".format(base, release_folder)
-
-
-def pick_largest_video(filenames, size_of):
-    """Return the largest video file from a list, or None.
-
-    ``size_of`` is a callable mapping filename -> size (bytes). Mirrors the
-    addon's "largest video wins" rule.
-    """
-    best = None
-    best_size = -1
-    for name in filenames:
-        lower = name.lower()
-        if not any(lower.endswith(ext) for ext in VIDEO_EXTENSIONS):
-            continue
-        size = size_of(name)
-        if size > best_size:
-            best = name
-            best_size = size
-    return best
-
-
-_SMB_LIST_RETRY_INTERVAL = 1.0
-# Total wall-clock budget to keep re-listing the SMB share after NZBGet
-# reports SUCCESS. NZBGet only enters history once its move is marked
-# complete, but the moved files can take a while longer to become listable
-# over the Samba export (observed: the containing folder's mtime landing at
-# the exact second we start looking). The old ~4s (5×1s) window lost that
-# race and failed with "No video file found on SMB share" even though the
-# download succeeded. 60s absorbs the visibility lag with wide margin.
-_SMB_RESOLVE_BUDGET = 60.0
-# Releases whose archive unpacks into a nested ``<release>/<inner>/video``
-# layout are common; descend a few levels (like the WebDAV resolver) so they
-# still resolve. Bounded to keep a pathological tree from stalling playback.
-_SMB_MAX_DEPTH = 3
-
-
-def _smb_file_size(path):
-    try:
-        return xbmcvfs.Stat(path).st_size()
-    except Exception:  # pylint: disable=broad-except
-        return 0
-
-
-def _is_video_name(name):
-    """True when ``name`` ends in one of the playable video extensions."""
-    lower = name.lower()
-    return any(lower.endswith(ext) for ext in VIDEO_EXTENSIONS)
-
-
-def _largest_video_in_dir(folder, files):
-    """Return ``(url, size)`` for the largest video file directly in ``folder``.
-
-    Considers only the given ``files`` (no descent). Returns ``(None, -1)``
-    when none are playable videos.
-    """
-    best, best_size = None, -1
-    for name in files:
-        if not _is_video_name(name):
-            continue
-        path = "{}/{}".format(folder, name)
-        size = _smb_file_size(path)
-        if size > best_size:
-            best, best_size = path, size
-    return best, best_size
-
-
-def _largest_video_in_tree(folder, depth=_SMB_MAX_DEPTH):
-    """Return ``(url, size)`` for the largest video at or below ``folder``.
-
-    Descends up to ``depth`` levels of subdirectories so a video tucked
-    inside a top-level folder (a common archive layout) still resolves
-    instead of failing with "No video file found on SMB share". Returns
-    ``(None, -1)`` when nothing playable is found.
-    """
-    try:
-        dirs, files = xbmcvfs.listdir(folder)
-    except Exception:  # pylint: disable=broad-except
-        return None, -1
-    best, best_size = _largest_video_in_dir(folder, files)
-    if depth > 0:
-        for sub in dirs:
-            url, size = _largest_video_in_tree("{}/{}".format(folder, sub), depth - 1)
-            if url is not None and size > best_size:
-                best, best_size = url, size
-    return best, best_size
-
-
-def _drive_resolve_dialog(dialog, now, deadline, budget):
-    """Advance the resolve progress bar; return True if the user canceled.
-
-    No-op (returns False) when no ``dialog`` is supplied. Otherwise drives the
-    progress bar over the resolve window so the user sees a "finishing up"
-    indicator instead of being dropped back to the home screen while the file
-    settles onto the share.
-    """
-    if dialog is None:
-        return False
-    if dialog.iscanceled():
-        return True
-    elapsed = budget - max(0.0, deadline - now)
-    percent = int(min(100.0, elapsed * 100.0 / budget)) if budget else 100
-    dialog.update(percent, _string(30219))
-    return False
-
-
-def resolve_smb_video(
-    smb_folder,
-    monitor=None,
-    dialog=None,
-    interval=_SMB_LIST_RETRY_INTERVAL,
-    budget=_SMB_RESOLVE_BUDGET,
-):
-    """List an SMB folder and return the largest video file URL, or None.
-
-    Searches the folder tree (top level plus nested subdirectories, see
-    ``_largest_video_in_tree``) and keeps retrying until a video appears or
-    the wall-clock ``budget`` (seconds, ``time.monotonic``) elapses — long
-    enough to absorb the lag between NZBGet reporting SUCCESS and the moved
-    files becoming visible over SMB. Sleeps ``interval`` seconds between
-    attempts via ``Monitor.waitForAbort`` so it stays cancelable and honors
-    Kodi shutdown. When a ``dialog`` (DialogProgress) is supplied, it shows a
-    progress bar over the wait and its Cancel button aborts the search.
-    Returns None if no video file appears within the budget.
-    """
-    if monitor is None:
-        monitor = xbmc.Monitor()
-    deadline = time.monotonic() + budget
-    while True:
-        url, _size = _largest_video_in_tree(smb_folder)
-        if url is not None:
-            return url
-        now = time.monotonic()
-        if now >= deadline:
-            return None
-        if _drive_resolve_dialog(dialog, now, deadline, budget):
-            return None
-        if monitor.waitForAbort(interval):
-            return None
-
-
 _POLL_INTERVAL = 2.0
 
 # listgroups Status values that mean the job is still actively downloading
@@ -694,373 +501,6 @@ def _build_submit_ctx(
     ctx.on_failure = on_failure
     ctx.dupe = dupe
     return ctx
-
-
-def _submit_dupe_backups(backups, dupe_key, settings_getter, cancel_event=None):
-    """Submit the release's duplicate backups to NZBGet (#372, Smart Duplicates).
-
-    ``backups`` is the picker-computed list of ``{"link","title","score"}`` for
-    the same-release-name reposts. Each is appended with the shared ``dupe_key``,
-    its own DupeScore (all below the pick's), and DupeMode=SCORE, so NZBGet keeps
-    the pick (highest score) downloading and parks each backup in history as a
-    duplicate -- failing over to the best remaining one if the pick is
-    unrepairable. Because NZBGet decides by score, submission order does not
-    matter: a backup submitted even after the pick has already succeeded is put
-    into history as a backup (not deleted). Best-effort: a bad/duplicate URL or a
-    failed fetch/append for one backup never aborts the rest or the pick. Stops
-    early if ``cancel_event`` fires (the user canceled the resolve). Returns the
-    list of submitted NZBIDs (for logging/tests).
-    """
-    submitted = []
-    seen = set()
-    index = 0
-    for backup in backups or []:
-        if cancel_event is not None and cancel_event.is_set():
-            break
-        nzb_url = _usable_backup_link(backup, seen)
-        if not nzb_url:
-            continue
-        seen.add(nzb_url)
-        index += 1
-        nzbid = _append_one_backup(nzb_url, backup, dupe_key, index, settings_getter)
-        if nzbid:
-            submitted.append(nzbid)
-    return submitted
-
-
-def _usable_backup_link(candidate, seen):
-    """The candidate's usable NZB link, or None to skip the row.
-
-    Shared filter for the picker's same-name backups and the loader-widened
-    extras: skip non-dict rows (defensive -- both lists are best-effort inputs)
-    and links already accepted this pass or already submitted (``seen``), so one
-    URL is never appended to NZBGet twice under the same DupeKey.
-    """
-    if not isinstance(candidate, dict):
-        return None
-    link = candidate.get("link")
-    if not link or link in seen:
-        return None
-    return link
-
-
-def _append_one_backup(nzb_url, backup, dupe_key, index, settings_getter):
-    """Append one duplicate backup to NZBGet and log the outcome (#372).
-
-    Returns the new NZBID, or None on a failed/raised append -- the caller keeps
-    iterating either way (best-effort: one bad backup never aborts the rest).
-    """
-    from resources.lib.fallback_streams import build_fallback_job_name
-
-    score = int(backup.get("score") or 0)
-    job_name = build_fallback_job_name(backup.get("title") or dupe_key, nzb_url, index)
-    try:
-        nzbid, error = nzbget_api.append_nzb(
-            nzb_url,
-            job_name,
-            settings_getter=settings_getter,
-            dupe_key=dupe_key,
-            dupe_score=score,
-            dupe_mode="SCORE",
-        )
-    except Exception as exc:  # pylint: disable=broad-except
-        xbmc.log(
-            "NZB-DAV: NZBGet duplicate backup submit raised: {}".format(
-                _redact_text(str(exc))
-            ),
-            xbmc.LOGWARNING,
-        )
-        return None
-    if nzbid:
-        xbmc.log(
-            "NZB-DAV: Queued NZBGet duplicate backup '{}' (score {})".format(
-                job_name, score
-            ),
-            xbmc.LOGINFO,
-        )
-        return nzbid
-    xbmc.log(
-        "NZB-DAV: NZBGet duplicate backup submit failed: {}".format(error),
-        xbmc.LOGINFO,
-    )
-    return None
-
-
-# Warn about HealthCheck=Pause at most once per Kodi session (a list so the
-# module-level flag is mutable from the worker thread; a lock so two concurrent
-# resolves' background threads can't both slip through the check-then-set).
-_HEALTHCHECK_WARNED = [False]
-_HEALTHCHECK_LOCK = threading.Lock()
-
-
-def _warn_if_healthcheck_pauses(settings_getter):
-    """Warn if NZBGet's ``HealthCheck=Pause`` disables automatic dup failover.
-
-    Per nzbget.com/documentation/rss/#duplicates automatic duplicate failover
-    needs HealthCheck = Delete, None (or Park); with Pause NZBGet pauses a failed
-    download instead of promoting a backup, so the picked release's backups sit
-    idle until the user unpauses one. Best-effort -- an unreadable config is
-    skipped. Always logs; notifies the user at most once per Kodi session.
-    """
-    try:
-        value = nzbget_api.config_option("HealthCheck", settings_getter=settings_getter)
-    except Exception:  # pylint: disable=broad-except
-        return
-    if value != "pause":
-        return
-    xbmc.log(
-        "NZB-DAV: NZBGet HealthCheck=Pause disables automatic duplicate failover; "
-        "set it to Delete or None to enable it (#372).",
-        xbmc.LOGWARNING,
-    )
-    with _HEALTHCHECK_LOCK:
-        if _HEALTHCHECK_WARNED[0]:
-            return
-        _HEALTHCHECK_WARNED[0] = True
-    _notify(_addon_name(), _string(30230), 6000)
-
-
-def _snapshot_conn_getter(settings_getter):
-    """A thread-safe getter over a main-thread snapshot of NZBGet connection
-    settings (#372).
-
-    Read the connection settings once on the calling (main/resolve) thread so the
-    background backup worker performs NO off-thread Kodi ``getSetting`` (unsafe on
-    CoreELEC/Kodi builds), and preserves a blank ``nzbget_username``/password
-    verbatim (``dict.get`` returns the stored ``""`` rather than the auth default
-    the addon getter substitutes).
-    """
-    url, user, password, category = nzbget_api._get_settings(settings_getter)
-    snapshot = {
-        "nzbget_url": url,
-        "nzbget_username": user,
-        "nzbget_password": password,
-        "nzbget_category": category,
-    }
-    return lambda key, default="": snapshot.get(key, default)
-
-
-def _dupe_check_disabled(settings_getter):
-    """True only when NZBGet's ``DupeCheck`` option is explicitly ``no``.
-
-    With DupeCheck off NZBGet does not park same-key items as backups -- it would
-    download every one as a normal queue item (parallel full downloads). Best-
-    effort: an unreadable config returns False (assume the default, on).
-    """
-    try:
-        return (
-            nzbget_api.config_option("DupeCheck", settings_getter=settings_getter)
-            == "no"
-        )
-    except Exception:  # pylint: disable=broad-except
-        return False
-
-
-_MAX_EXTRA_BACKUPS = 5
-
-
-def _extra_backups_from_loader(loader, seen_links, limit=_MAX_EXTRA_BACKUPS):
-    """Same-content / NZBHydra-deferred candidates from the fallback loader.
-
-    #372 round 2 widening: beyond the picker's exact same-name rows, the fallback
-    loader (an indexer search, already threaded for the nzbdav path) surfaces the
-    same-content mirrors and NZBHydra duplicate uploads that were collapsed into a
-    single picker row. Returns ``[{"link","title","score"}]`` deduped against
-    ``seen_links``, scored DESCENDING from 0 so they sit BELOW every same-name
-    backup (a last-resort failover, keyed under the pick's DupeKey). Bounded by
-    ``limit`` (the standby cap's remaining slots, hard-capped at
-    ``_MAX_EXTRA_BACKUPS``) so the total backup count honors the user's
-    "Maximum standby fallback streams". Best-effort: a missing/erroring loader,
-    its "disabled" sentinel (a non-list), or ``limit <= 0`` yields ``[]``.
-    """
-    cap = min(limit, _MAX_EXTRA_BACKUPS)
-    if loader is None or cap <= 0:
-        return []
-    extras = []
-    seen = set(seen_links or [])
-    score = 0
-    for candidate in _load_extra_candidates(loader):
-        if len(extras) >= cap:
-            break
-        link = _usable_backup_link(candidate, seen)
-        if not link:
-            continue
-        seen.add(link)
-        extras.append({"link": link, "title": candidate.get("title"), "score": score})
-        score -= 1
-    return extras
-
-
-def _load_extra_candidates(loader):
-    """Run the fallback loader, absorbing every failure mode (#372 r2).
-
-    Returns the candidate list, or ``[]`` for an erroring loader or its
-    "disabled" sentinel (a non-list) -- the extras are best-effort widening
-    only, so a broken indexer search must never surface past here.
-    """
-    try:
-        candidates = loader()
-    except Exception:  # pylint: disable=broad-except
-        return []
-    return candidates if isinstance(candidates, list) else []
-
-
-def _dupe_worker_should_skip(getter, cancel_event):
-    """True when the backup worker must submit nothing (#372).
-
-    Skips silently on a pre-submit cancel (the user already gave up on the
-    resolve), and skips with a log when the server has DupeCheck=no -- same-key
-    items would then download in parallel instead of parking as backups.
-    """
-    if cancel_event.is_set():
-        return True
-    if _dupe_check_disabled(getter):
-        xbmc.log(
-            "NZB-DAV: NZBGet DupeCheck=no -- skipping #372 duplicate "
-            "backups (they would download in parallel).",
-            xbmc.LOGINFO,
-        )
-        return True
-    return False
-
-
-def _submit_backup_fleet(
-    getter, cancel_event, dupe_key, backups, loader, extras_limit, submitted_ids
-):
-    """Submit the same-name backups, then the loader-widened extras (#372).
-
-    Widens with same-content / Hydra-deferred candidates (#372 r2) as
-    lowest-priority backups keyed under the same (pick's) DupeKey. Bounds them
-    by the standby cap's REMAINING slots so same-name backups + extras never
-    exceed "Maximum standby fallback streams". Reads ONLY the snapshot
-    ``getter`` -- this runs on the worker thread, which must never touch Kodi.
-    Every appended NZBID is recorded into ``submitted_ids`` AS IT LANDS so the
-    post-cancel cleanup can delete exactly this resolve's submissions.
-    """
-    submitted_ids.extend(
-        _submit_dupe_backups(backups, dupe_key, getter, cancel_event=cancel_event) or []
-    )
-    if cancel_event.is_set():
-        return
-    remaining = (
-        _MAX_EXTRA_BACKUPS
-        if extras_limit is None
-        else max(0, extras_limit - len(backups))
-    )
-    extras = _extra_backups_from_loader(
-        loader, [b.get("link") for b in backups], limit=remaining
-    )
-    if extras:
-        submitted_ids.extend(
-            _submit_dupe_backups(extras, dupe_key, getter, cancel_event=cancel_event)
-            or []
-        )
-
-
-def _cleanup_canceled_submissions(getter, submitted_ids):
-    """Delete exactly THIS worker's submissions after a mid-submit cancel (#372).
-
-    Covers the append that was already in flight when the user canceled and so
-    landed after _handle_poll_failure's one-shot DupeKey sweep. Scoped to the
-    NZBIDs this resolve submitted -- NEVER a whole-DupeKey sweep, which could
-    wipe a fresh retry of the same release (it shares the stable DupeKey) that
-    started while this stale worker drained. Best-effort like every other
-    backup step: a failed delete is swallowed, never raised off-thread.
-    """
-    try:
-        nzbget_api.cancel_jobs(submitted_ids, settings_getter=getter)
-    except Exception:  # pylint: disable=broad-except
-        pass
-
-
-def _spawn_dupe_backups(ctx):
-    """Fire-and-forget the release's duplicate backups in a daemon thread (#372).
-
-    Runs off the resolve thread (each backup is an indexer HTTP round-trip) so it
-    never delays the pick's poll/progress ("it won't affect playback"); the
-    daemon flag keeps it from blocking Kodi shutdown. Because every item carries
-    an explicit DupeScore (the pick highest), NZBGet keeps the pick the active
-    download regardless of when the backups land -- so submission order is not a
-    concern and a backup arriving after the pick already succeeded is still put
-    into history as a backup, not deleted. Skips entirely when the server has
-    DupeCheck disabled (backups would download in parallel), and warns once if
-    HealthCheck=Pause would block automatic failover. Reads settings from a
-    main-thread snapshot so the worker never touches Kodi off-thread. All errors
-    are swallowed -- backups are pure insurance and must never break playback.
-    """
-    dupe = ctx.dupe or {}
-    dupe_key = dupe.get("key") or ""
-    backups = list(dupe.get("backups") or [])
-    if not dupe_key or not backups:
-        return None
-    try:
-        getter = _snapshot_conn_getter(ctx.settings_getter)
-    except Exception as exc:  # pylint: disable=broad-except
-        # The snapshot reads Kodi/injected settings and runs AFTER the primary is
-        # already accepted. Backups are pure insurance -- a settings-read failure
-        # here must skip them, never propagate out and fail the primary's playback.
-        xbmc.log(
-            "NZB-DAV: NZBGet duplicate backup snapshot failed: {}".format(
-                _redact_text(str(exc))
-            ),
-            xbmc.LOGWARNING,
-        )
-        return None
-    cancel_event = ctx.cancel_event
-    loader = dupe.get("loader")
-    extras_limit = dupe.get("max_backups")
-
-    def _worker():
-        reached_submit = False
-        submitted_ids = []
-        try:
-            if _dupe_worker_should_skip(getter, cancel_event):
-                return
-            _warn_if_healthcheck_pauses(getter)
-            reached_submit = True
-            _submit_backup_fleet(
-                getter,
-                cancel_event,
-                dupe_key,
-                backups,
-                loader,
-                extras_limit,
-                submitted_ids,
-            )
-        except Exception as exc:  # pylint: disable=broad-except
-            xbmc.log(
-                "NZB-DAV: NZBGet duplicate backup worker error: {}".format(
-                    _redact_text(str(exc))
-                ),
-                xbmc.LOGWARNING,
-            )
-        finally:
-            # If a cancel arrived while a backup's append was already in flight,
-            # that backup can land in NZBGet AFTER _handle_poll_failure's
-            # one-shot cancel_dupekey_group sweep -- and NZBGet would then
-            # promote the orphan as the group's new active download. Clean up
-            # once the worker has drained, scoped to this resolve's own
-            # submissions (#372 r2 cancel-race, r3 retry-race).
-            if reached_submit and cancel_event.is_set() and submitted_ids:
-                _cleanup_canceled_submissions(getter, submitted_ids)
-
-    try:
-        thread = threading.Thread(
-            target=_worker, name="nzbdav-nzbget-dupe-backups", daemon=True
-        )
-        thread.start()
-    except Exception as exc:  # pylint: disable=broad-except
-        # e.g. RuntimeError "can't start new thread" under thread exhaustion.
-        # The backups are pure insurance -- never let them break the already-
-        # queued pick's playback.
-        xbmc.log(
-            "NZB-DAV: NZBGet duplicate backup spawn failed: {}".format(
-                _redact_text(str(exc))
-            ),
-            xbmc.LOGWARNING,
-        )
-        return None
-    return thread
 
 
 def _submit_poll_resolve(ctx, nzb_url, title, download_pubdate, download_size):
